requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
audioread==3.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
editdistance==0.8.1
eng_to_ipa==0.0.2
epitran==1.26.0
exceptiongroup==1.2.2
executing==2.2.0
fireworks-ai==0.15.12
flake8==7.1.1
fonttools==4.56.0
grpcio==1.70.0
grpcio-health-checking==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jamo==0.4.1
jedi==0.19.2
jiter==0.8.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.26.1
librosa==0.10.2.post1
llvmlite==0.44.0
marisa-trie==1.2.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
msgpack==1.1.0
munkres==1.1.4
nest-asyncio==1.6.0
nltk==3.9.1
numba==0.61.0
numpy==2.1.3
openai==1.63.0
packaging==24.2
panphon==0.21.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
praat-parselmouth==0.4.5
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.2.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.12.1
regex==2024.11.6
requests==2.32.3
ruff==0.9.6
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
stack-data==0.6.3
syllapy==0.7.2
textblob==0.19.0
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
unicodecsv==0.14.1
urllib3==2.3.0
wcwidth==0.2.13
wsproto==1.2.0
